+++ Source/Streamstone.Tests/Scenarios/Writing_to_stream.cs
@@ -321,7 +321,7 @@ namespace Streamstone.Scenarios
             AssertStreamEntity(3, replaced);
         }
 
-        void AssertNewStream(StreamWriteResult actual, int version, object properties = null)
+        void AssertNewStream(StreamWriteResult actual, long version, object properties = null)
         {
             var newStream = actual.Stream;
             var newStreamEntity = partition.RetrieveStreamEntity();
@@ -330,7 +330,7 @@ namespace Streamstone.Scenarios
             expectedStream.ToExpectedObject().ShouldEqual(newStream);
         }
 
-        void AssertModifiedStream(Stream previous, StreamWriteResult actual, int version)
+        void AssertModifiedStream(Stream previous, StreamWriteResult actual, long version)
         {
             var actualStream = actual.Stream;
             var actualStreamEntity = partition.RetrieveStreamEntity();
@@ -341,7 +341,7 @@ namespace Streamstone.Scenarios
             expectedStream.ToExpectedObject().ShouldEqual(actualStream);
         }
 
-        Stream CreateStream(int version, ETag etag, object properties = null)
+        Stream CreateStream(long version, ETag etag, object properties = null)
         {
             var props = properties != null
                 ? StreamProperties.From(properties)
@@ -350,7 +350,7 @@ namespace Streamstone.Scenarios
             return new Stream(partition, etag, version, props);
         }
 
-        void AssertStreamEntity(int version = 0, object properties = null)
+        void AssertStreamEntity(long version = 0, object properties = null)
         {
             var newStreamEntity = partition.RetrieveStreamEntity();
 
@@ -366,13 +366,13 @@ namespace Streamstone.Scenarios
             expectedEntity.ToExpectedObject().ShouldMatch(newStreamEntity);
         }
 
-        void AssertRecordedEvent(int version, EventData source, RecordedEvent actual)
+        void AssertRecordedEvent(long version, EventData source, RecordedEvent actual)
         {
             var expected = source.Record(partition, version);
             expected.ToExpectedObject().ShouldMatch(actual);
         }
 
-        static void AssertEventEntity(int version, EventEntity actual)
+        static void AssertEventEntity(long version, EventEntity actual)
         {
             var expected = new
             {
@@ -389,7 +389,7 @@ namespace Streamstone.Scenarios
             Assert.That(actual.ETag.ToString, Is.Not.Empty);
         }
 
-        static void AssertEventIdEntity(string id, int version, EventIdEntity actual)
+        static void AssertEventIdEntity(string id, long version, EventIdEntity actual)
         {
             var expected = new
             {
